# swagger.yml
swagger: "2.0"
info:
  description: "This is an API for creating maintenance tasks for factory devices."
  version: "1.0.0"
  title: "Service Manual API"
  contact:
    name: "Henri Hakkarainen"
    email: "hakkarainenhenri@gmail.com"
host: "localhost:3000"
basePath: "/api"
tags:
- name: "task"
  description: "Operations for maintenance tasks"
schemes:
- "http"
paths:
  /tasks:
    post:
      tags:
      - "task"
      summary: "Add a new task to the database"
      description: ""
      operationId: "addTask"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Task object that is going to be created"
        required: true
        schema:
          type: "object"
          properties:
            description:
              type: "string"
            priority:
              type: "string"
            mode:
              type: "string"
            deviceid:
              type: "integer"
              format: "int64"
      responses:
        "201":
          description: "Operation successful"
          schema:
            type: "object"
            properties:
              status:
                type: "string"
              data:
                $ref: "#/definitions/Task"
        "400":
          description: "Invalid input"
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: "Operation failed"
          schema:
            $ref: "#/definitions/ErrorResponse"
    get:
      tags:
      - "task"
      summary: "Receive a list of all tasks or just tasks of a single device"
      description: "Query parameter deviceId can be provided to filter the results to return a list of tasks of a single device instead"
      operationId: "getTasks"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "deviceId"
        in: "query"
        description: "DeviceID that is to be considered for filtering results"
        required: false
        type: "integer"
      responses:
        "200":
          description: "Successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Task"
        "400":
          description: "Invalid deviceId value provided"
          schema:
            $ref: "#/definitions/ErrorResponse"
        "404":
          description: "Device matching query parameter was not found"
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: "Operation failed"
          schema:
            $ref: "#/definitions/ErrorResponse"
  /tasks/{taskId}:
    get:
      tags:
      - "task"
      summary: "Find a specific task by ID"
      description: "Returns a single task"
      operationId: "getTaskById"
      produces:
      - "application/json"
      parameters:
      - name: "taskId"
        in: "path"
        description: "ID of task to return"
        required: true
        type: "integer"
      responses:
        "200":
          description: "Operation successful"
          schema:
            $ref: "#/definitions/Task"
        "404":
          description: "Task not found"
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: "Operation failed"
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      tags:
      - "task"
      summary: "Updates an existing task"
      description: ""
      operationId: "updateTaskById"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "taskId"
        in: "path"
        description: "ID of task that needs to be updated"
        required: true
        type: "integer"
      - name: "body"
        in: "body"
        description: "Updated task object"
        required: true
        schema:
          type: "object"
          properties:
            description:
              type: "string"
            priority:
              type: "string"
            mode:
              type: "string"
      responses:
        "200":
          description: "Operation successful"
          schema:
            type: "object"
            properties:
              status:
                type: "string"
              data:
                $ref: "#/definitions/Task"
        "204":
          description: "No request body provided so nothing happens"
        "400":
          description: "Invalid input values"
          schema:
            $ref: "#/definitions/ErrorResponse"
        "404":
          description: "Task not found"
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: "Operation failed"
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      tags:
      - "task"
      summary: "Deletes a single task"
      description: ""
      operationId: "deleteTask"
      produces:
      - "application/json"
      parameters:
      - in: "path"
        name: "taskId"
        description: "ID of the task that is to be deleted"
        required: true
        type: "integer"
      responses:
        "200":
          description: "Task deleted"
          schema:
            type: "object"
            properties:
              status:
                type: "string"
              data:
                $ref: "#/definitions/Task"
        "404":
          description: "Task not found"
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: "Operation failed"
          schema:
            $ref: "#/definitions/ErrorResponse"

definitions:
  Task:
    type: "object"
    properties:
      taskid:
        type: "integer"
        format: "int64"
        description: "Unique identifier of the task"
      entrydate:
        type: "string"
        description: "Date of entry"
      description:
        type: "string"
        description: "Description of the task"
      priority:
        type: "string"
        description: "Priority of the task"
        enum:
        - "critical"
        - "important"
        - "slight"
      mode:
        type: "string"
        description: "Current state of the task"
        enum:
        - "open"
        - "done"
      deviceid:
        type: "integer"
        format: "int64"
        description: "ID of the device the task is pointed at"
  ErrorResponse:
    type: "object"
    properties:
      status:
        type: "string"
      error:
        type: "string"
        description: "Error message from the server"